thesis_offline_validation/requirements.txt
numpy
pandas
pyarrow
opencv-python
PyYAML
ultralytics

